Fix v20 training

+++ lerobot/common/datasets/v21/convert_dataset_v20_to_v21.py
@@ -55,11 +55,14 @@ class SuppressWarnings:
 
 def convert_dataset(
     repo_id: str,
+    root: str | None = None,
     branch: str | None = None,
     num_workers: int = 4,
+    sync : bool = False,
+    upload: bool = False
 ):
     with SuppressWarnings():
-        dataset = LeRobotDataset(repo_id, revision=V20, force_cache_sync=True)
+        dataset = LeRobotDataset(repo_id, root=root, revision=V20, force_cache_sync=sync)
 
     if (dataset.root / EPISODES_STATS_PATH).is_file():
         (dataset.root / EPISODES_STATS_PATH).unlink()
@@ -71,21 +74,23 @@ def convert_dataset(
     dataset.meta.info["codebase_version"] = CODEBASE_VERSION
     write_info(dataset.meta.info, dataset.root)
 
-    dataset.push_to_hub(branch=branch, tag_version=False, allow_patterns="meta/")
+    if upload:
+        dataset.push_to_hub(branch=branch, tag_version=False, allow_patterns="meta/")
 
     # delete old stats.json file
     if (dataset.root / STATS_PATH).is_file:
         (dataset.root / STATS_PATH).unlink()
 
-    hub_api = HfApi()
-    if hub_api.file_exists(
-        repo_id=dataset.repo_id, filename=STATS_PATH, revision=branch, repo_type="dataset"
-    ):
-        hub_api.delete_file(
-            path_in_repo=STATS_PATH, repo_id=dataset.repo_id, revision=branch, repo_type="dataset"
-        )
+    if upload:
+        hub_api = HfApi()
+        if hub_api.file_exists(
+            repo_id=dataset.repo_id, filename=STATS_PATH, revision=branch, repo_type="dataset"
+        ):
+            hub_api.delete_file(
+                path_in_repo=STATS_PATH, repo_id=dataset.repo_id, revision=branch, repo_type="dataset"
+            )
 
-    hub_api.create_tag(repo_id, tag=CODEBASE_VERSION, revision=branch, repo_type="dataset")
+        hub_api.create_tag(repo_id, tag=CODEBASE_VERSION, revision=branch, repo_type="dataset")
 
 
 if __name__ == "__main__":
@@ -108,6 +113,12 @@ if __name__ == "__main__":
         type=int,
         default=4,
         help="Number of workers for parallelizing stats compute. Defaults to 4.",
+    )
+    parser.add_argument(
+        "--root",
+        type=str,
+        default=None,
+        help="local root data dir.",
     )
 
     args = parser.parse_args()
